Use fused mul-add instructions where possible
The Rust compiler does not yet optimize FMA instructions in a
majority of cases. Therefore, it is recommended to use the `f32::mul_add`
method to allow FMA instructions to be used. In some cases this can
provide a significant speedup on machines with FMA available,
and the fused mul-add instruction is reported to be more accurate
than a manual floating point mul and add instruction.